fix ios to match new heimdall
