+++ experimental/mobile/examples/ios-nasp/ios-nasp.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "self:/Users/nandork/Code/src/github.com/cisco-open/nasp/experimental/mobile/examples/ios-nasp/ios-nasp.xcodeproj">
+   </FileRef>
+</Workspace>
