+++ experimental/mobile/examples/ios-nasp/ios-nasp.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
