+++ experimental/mobile/examples/ios-nasp/ios-nasp/Assets.xcassets/AccentColor.colorset/Contents.json
