+++ experimental/mobile/examples/ios-nasp/ios-nasp/Assets.xcassets/AppIcon.appiconset/Contents.json
