+++ experimental/mobile/examples/ios-nasp/ios-nasp/Assets.xcassets/Contents.json
