+++ experimental/mobile/examples/ios-nasp/ios-naspUITests/ios_istioUITests.swift
@@ -13,12 +13,16 @@
 //  limitations under the License.
 
 import XCTest
-@testable import ios_istio
 
-class ios_istioTests: XCTestCase {
+class ios_naspUITests: XCTestCase {
 
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
+
+        // In UI tests it is usually best to stop immediately when a failure occurs.
+        continueAfterFailure = false
+
+        // In UI tests it’s important to set the initial state - such as interface orientation - required for your tests before they run. The setUp method is a good place to do this.
     }
 
     override func tearDownWithError() throws {
@@ -26,18 +30,19 @@ class ios_istioTests: XCTestCase {
     }
 
     func testExample() throws {
-        // This is an example of a functional test case.
+        // UI tests must launch the application that they test.
+        let app = XCUIApplication()
+        app.launch()
+
         // Use XCTAssert and related functions to verify your tests produce the correct results.
-        // Any test you write for XCTest can be annotated as throws and async.
-        // Mark your test throws to produce an unexpected failure when your test encounters an uncaught error.
-        // Mark your test async to allow awaiting for asynchronous code to complete. Check the results with assertions afterwards.
     }
 
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
+    func testLaunchPerformance() throws {
+        if #available(macOS 10.15, iOS 13.0, tvOS 13.0, watchOS 7.0, *) {
+            // This measures how long it takes to launch your application.
+            measure(metrics: [XCTApplicationLaunchMetric()]) {
+                XCUIApplication().launch()
+            }
         }
     }
-
 }
